e2e colocation clean-up
