+++ crates/e2e/tests/e2e/main.rs
@@ -6,25 +6,22 @@
 // Each of the following modules contains tests.
 mod app_data;
 mod app_data_signer;
-mod colocation_buffers;
-mod colocation_eth_safe;
-mod colocation_ethflow;
-mod colocation_hooks;
-mod colocation_partial_fill;
-mod colocation_quoting;
-mod colocation_solver_competition;
-mod colocation_univ2;
+mod buffers;
 mod database;
 mod eth_integration;
+mod eth_safe;
 mod ethflow;
 mod hooks;
 mod limit_orders;
 mod onchain_settlement;
 mod order_cancellation;
+mod partial_fill;
 mod partially_fillable_balance;
 mod partially_fillable_pool;
 mod quoting;
 mod refunder;
 mod smart_contract_orders;
+mod solver_competition;
 mod tracking_insufficient_funds;
+mod univ2;
 mod vault_balances;
