+++ crates/e2e/tests/e2e/partial_fill.rs
@@ -1,6 +1,6 @@
 use {
     e2e::{
-        setup::{colocation::SolverEngine, *},
+        setup::{*},
         tx,
         tx_value,
     },
@@ -45,22 +45,8 @@ async fn test(web3: Web3) {
     );
 
     tracing::info!("Starting services.");
-    let solver_endpoint =
-        colocation::start_baseline_solver(onchain.contracts().weth.address()).await;
-    colocation::start_driver(
-        onchain.contracts(),
-        vec![SolverEngine {
-            name: "test_solver".into(),
-            account: solver,
-            endpoint: solver_endpoint,
-        }],
-    );
-
     let services = Services::new(onchain.contracts()).await;
-    services.start_autopilot(vec![
-        "--drivers=test_solver|http://localhost:11088/test_solver".to_string(),
-    ]);
-    services.start_api(vec![]).await;
+    services.start_protocol(solver).await;
 
     tracing::info!("Placing order");
     let balance = token.balance_of(trader.address()).call().await.unwrap();
