+++ crates/e2e/tests/e2e/solver_competition.rs
